requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.35.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.17
redis==5.0.1 